requirements.txt
# 参照Plot Ark的成功配置
Flask==2.3.3
Werkzeug==2.3.7
gunicorn
python-dotenv
google-generativeai
Flask-Cors
